RESTRUCTURED THE PROJECT
